LINUX ADMINISTRATION 8TH ASSIGNMENT
Assignment 0
1. Create a simple shell script to tell the user about their session – they
 need to know:
● What their username is
● What the current date is
● What the time is
● What their current working directory is
● How many files they have in that directory
● What is the biggest file in their current directory
Assignment 1
Create a directory with a few test files in it (the files can be empty). Now write a script that for
every file in that directory you rename it to have an extension of today's date in YYYYMMDD
format.

Assignment 2
Write a script that takes a number as an input and reverses it out to the user. For example, if the
original number is 74985, the output should be 58947.

Assignment 3
Write a script to validate how secure someone's password is. Things you would care about:
● Length should be 8 or more characters
● The password should contain numbers and letters
● There should be both uppercase and lowercase letters